requirements.txt
asgiref==3.3.1
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
cssselect2==0.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.43.0
djangorestframework==3.12.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
oauthlib==3.1.0
Pillow==8.0.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
Pyphen==0.10.0
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
urllib3==1.26.0
WeasyPrint==52.1
webencodings==0.5.1
whitenoise==5.2.0
